% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculo.probabilidad.R
\name{calculo.probabilidad}
\alias{calculo.probabilidad}
\title{Cálculo de probabilidades.}
\usage{
calculo.probabilidad()
}
\value{
No devuelve un valor, es una aplicación shiny.
}
\description{
Aplicación interactiva para practicar el cálculo de probabilidades a partir de distintas distribuciones de probabilidad.
}
\author{
\strong{Vicente Coll-Serrano}.
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

\strong{Rosario Martínez Verdú}.
\emph{Economía Aplicada.}

Facultad de Economía. Universidad de Valencia (España)
}
