% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeficiente.variacion.R
\name{coeficiente.variacion}
\alias{coeficiente.variacion}
\title{Coeficiente de variación.}
\usage{
coeficiente.variacion(x, variable = NULL, pesos = NULL)
}
\arguments{
\item{x}{Conjunto de datos. Puede ser un vector o un dataframe.}

\item{variable}{Es un vector (numérico o carácter) que indica las variables a seleccionar de x. Si x se refiere una sola variable, el argumento variable es NULL. En caso contrario, es necesario indicar el nombre o posición (número de columna) de la variable.}

\item{pesos}{Si los datos de la variable están resumidos en una distribución de frecuencias, debe indicarse la columna que representa los valores de la variable y la columna con las frecuencias o pesos.}
}
\description{
Calcula el coeficiente de variación de Pearson.
}
\references{
Esteban García, J. et al. (2005). Estadística descriptiva y nociones de probabilidad. Thomson.
}
\author{
\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

\strong{Olga Blasco-Blasco} (\email{olga.blasco@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

\strong{Rosario Martínez Verdú} (\email{rosario.martinez@uv.es}).
\emph{Economía Aplicada.}

\strong{Cristina Pardo García} (\email{cristina.pardo-garcia@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

Universidad de Valencia (España)
}
