% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contraste.varianza.R
\name{contraste.varianza}
\alias{contraste.varianza}
\title{Contraste de hipótesis sobre la varianza.}
\usage{
contraste.varianza(x,
                       variable = NULL,
                       introducir = FALSE,
                       media_poblacion = c("desconocida","conocida"),
                       hipotesis_nula = NULL,
                       tipo_contraste =  c("bilateral","cola derecha","cola izquierda"),
                       alfa = 0.05,
                       grafico = FALSE))
}
\arguments{
\item{x}{Conjunto de datos. Puede ser un vector o un dataframe.}

\item{variable}{Es un vector (numérico o carácter) que indica las variables a seleccionar de x. Si x se refiere una sola variable, el argumento variable es NULL. En caso contrario, es necesario indicar el nombre o posición (número de columna) de la variable.}

\item{introducir}{Valor lógico. Si introducir = FALSE (por defecto), el usuario debe indicar el conjunto de datos que desea analizar usando los argumentos x y/o variable. Si introducir = TRUE, se le solicitará al ususario que introduzca la información relevante sobre tamaño muestral, valor de la media muestral, etc.}

\item{media_poblacion}{Es un carácter. Indica si la media de la población es desconocida (por defecto, media_poblacion = "desconocida") o conocida (en este caso, cambiar media_poblacion = "conocida").}

\item{hipotesis_nula}{Es un valor numérico.}

\item{tipo_contraste}{Es un carácter. Indica el tipo de contraste a realizar. Por defecto, tipo_contraste = "bilateral".
Si tipo_contraste = "bilateral", se contraste la hipótesis nula igual un valor frente a la alternativa distinto de dicho valor.
Si tipo_contraste = "cola derecha", se contrasta la hipótesis nula menor o igual a un valor frente a la alternativa mayor a dicho valor.
Si tipo_contraste = "cola izquierda", se contrasta la hipótesis nula mayor o igual a un valor frente a la alternativa menos a dicho valor.}

\item{alfa}{Es un valor numérico entre 0 y 1. Indica el nivel de significación. Por defecto, alfa = 0.05 (5 por ciento)}

\item{grafico}{Es un valor lógico. Por defecto grafico = FALSE. Si se quiere obtener una representación gráfica del intervalo de confianza obtenido, cambiar el argumento a grafico = TRUE. Nota: Esta opción no está implementada para todos los casos.}
}
\description{
Realiza el contraste de hipótesis sobre la varianza poblacional.
}
\references{
Esteban García, J. et al. (2005). Estadística descriptiva y nociones de probabilidad. Thomson.
}
\author{
\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

\strong{Olga Blasco-Blasco} (\email{olga.blasco@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

\strong{Rosario Martínez Verdú} (\email{rosario.martinez@uv.es}).
\emph{Economía Aplicada.}

\strong{Cristina Pardo García} (\email{cristina.pardo-garcia@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

Universidad de Valencia (España)
}
