% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergencia.varianza.R
\name{convergencia.varianza}
\alias{convergencia.varianza}
\title{Convergencia de la varianza y cuasivarianza muestral.}
\usage{
convergencia.varianza()
}
\value{
Devuelve un gráfico que es un objeto de la clase \code{plotly} y \code{htmlwidget}.
}
\description{
Gráfico dinámico que ilustra la convergencia de la varianza y cuasi-varianza muestral a medida que aumenta el tamaño muestral.
}
\author{
\strong{Vicente Coll-Serrano}.
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

\strong{Rosario Martínez Verdú}.
\emph{Economía Aplicada.}

\strong{Cristina Pardo-García}.
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

Facultad de Economía. Universidad de Valencia (España)
}
