% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.sets.R
\docType{data}
\name{diseno1}
\alias{diseno1}
\title{Datos simulados de dos muestras tomadas en periodos de tiempo distintos. La muestra 1 es tomada en enero y la muestra 2 en junio.}
\format{
Dataframe en formato ancho con 620 observaciones. La pregunta realizada es: ¿Sabe que Valencia es la capital mundial del diseño 2022?
\describe{
  \item{muestra1}{0: No sabe, 1: Sí Sabe}
  \item{muestra2}{0: No sabe, 1: Sí sabe}
}
}
\source{
Muestra simulada.
}
\usage{
data("diseno1")
}
\description{
Datos simulados de dos muestras tomadas en periodos de tiempo distintos. La muestra 1 es tomada en enero y la muestra 2 en junio.
}
\author{
\strong{Vicente Coll-Serrano}.
\emph{Quantitative Methods for Measuring Culture (MC2). Applied Economics.}

\strong{Rosario Martínez Verdú}.
\emph{Economía Aplicada.}

\strong{Cristina Pardo-García}.
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

Facultad de Economía. Universidad de Valencia (España)
}
\keyword{datasets}
