% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribucion.normal.R
\name{distribucion.normal}
\alias{distribucion.normal}
\title{Distribución normal.}
\usage{
distribucion.normal()
}
\value{
No devuelve un valor, es una aplicación shiny.
}
\description{
Aplicación interactiva para comparar dos distribuciones normales.
}
\author{
\strong{Vicente Coll-Serrano}.
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

\strong{Rosario Martínez Verdú}.
\emph{Economía Aplicada.}

\strong{Cristina Pardo-García}.
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

Facultad de Economía. Universidad de Valencia (España)
}
