% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribuciones.probabilidad.R
\name{distribuciones.probabilidad}
\alias{distribuciones.probabilidad}
\title{Distribuciones de probabilidad.}
\usage{
distribuciones.probabilidad()
}
\value{
No devuelve un valor, Es una aplicación shiny.
}
\description{
Aplicación interactiva donde se representa las principales distribuciones de probabilidad unidimensionales: Binomial, Poisson, Uniforme, Exponencial y Normal.
}
\author{
\strong{Vicente Coll-Serrano}.
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

\strong{Rosario Martínez Verdú}.
\emph{Economía Aplicada.}

\strong{Cristina Pardo-García}.
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

Facultad de Economía. Universidad de Valencia (España)
}
