% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.sets.R
\docType{data}
\name{ejem_bidi}
\alias{ejem_bidi}
\title{Data: Ejemplo de dos variables (ejem_bidi)}
\format{
Dataframe con 100 observaciones de 2 variables.
\describe{
  \item{x}{Toma valores de 0 a 5.}
  \item{x}{Toma valores de 10 a 15}
}
}
\source{
Muestra simulada.
}
\usage{
data("ejem_bidi")
}
\description{
Datos simulados. Muestra de 100 observaciones
}
\author{
\strong{Vicente Coll-Serrano}.
\emph{Quantitative Methods for Measuring Culture (MC2). Applied Economics.}

\strong{Rosario Martínez Verdú}.
\emph{Economía Aplicada.}

\strong{Cristina Pardo-García}.
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

Facultad de Economía. Universidad de Valencia (España)
}
\keyword{datasets}
