% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.sets.R
\docType{data}
\name{hogares}
\alias{hogares}
\title{Data: Hogares}
\format{
Dataframe con 10 observaciones de 3 variables.
\describe{
  \item{Hogares}{Identificación del hogar.}
  \item{ingresos}{Ingresos del hogar}
  \item{viajes}{Número de hogares realizado por los hogares.}
}
}
\usage{
data("hogares")
}
\description{
Datos de 10 hogares que se utilizan en los ejemplos de (1) tabla bidimensional, (2) covarianza, (3) matriz de covarianzas, (4) correlación y (5) matriz de correlación.
}
\keyword{datasets}
