% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.varianza.R
\name{ic.varianza}
\alias{ic.varianza}
\title{Intervalo confianza para la varianza.}
\usage{
ic.varianza(x,
           variable = NULL,
           introducir = FALSE,
           media_poblacion = c("desconocida","conocida"),
           confianza = 0.95)
}
\arguments{
\item{x}{Conjunto de datos. Puede ser un vector o un dataframe.}

\item{variable}{Es un vector (numérico o carácter) que indica las variables a seleccionar de x. Si x se refiere una sola variable, el argumento variable es NULL. En caso contrario, es necesario indicar el nombre o posición (número de columna) de la variable.}

\item{introducir}{Valor lógico. Si introducir = FALSE (por defecto), el usuario debe indicar el conjunto de datos que desea analizar usando los argumentos x y/o variable. Si introducir = TRUE, se le solicitará al ususario que introduzca la información relevante sobre tamaño muestral, valor de la media muestral, etc.}

\item{media_poblacion}{Es un carácter. Indica si la media de la población es desconocida (por defecto, media_poblacion = "desconocida") o conocida (en este caso, cambiar media_poblacion = "conocida").}

\item{confianza}{Es un valor numérico entre 0 y 1. Indica el nivel de confianza. Por defecto, confianza = 0.95 (95 por ciento)}

\item{var_muestra}{Es un carácter. Indica si se utilizará la varianza muestral (por defecto, var_muestra = "muestral") o la cuasivarianza muestral (es este caso, cambiar var_muestra = "cuasi").}
}
\description{
Calcula el intervalo de confianza de la varianza poblacional.
}
\references{
Esteban García, J. et al. (2005). Estadística descriptiva y nociones de probabilidad. Thomson.
}
\author{
\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

\strong{Olga Blasco-Blasco} (\email{olga.blasco@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

\strong{Rosario Martínez Verdú} (\email{rosario.martinez@uv.es}).
\emph{Economía Aplicada.}

\strong{Cristina Pardo García} (\email{cristina.pardo-garcia@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

Universidad de Valencia (España)
}
