% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer.datos.R
\name{leer.datos}
\alias{leer.datos}
\title{Leer datos.}
\usage{
leer.datos(introducir = FALSE, pos = 1)
}
\arguments{
\item{introducir}{Valor lógico. Si introducir = FALSE (por defecto), se abrirá una ventana para que el usuario seleccione el fichero de datos que quiere cargar. Si introducir = TRUE, el usuario introducirá él mismo los datos.}

\item{pos}{Es un valor fijo utilizado para mostrar el dataframe del usuario en el Global Environment.}
}
\description{
Carga un conjunto de datos.
}
\references{
Esteban García, J. et al. (2005). Estadística descriptiva y nociones de probabilidad. Thomson.
}
\author{
\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

\strong{Rosario Martínez Verdú} (\email{Rosario.Martinez@uv.es}).
\emph{Economía Aplicada.}

\strong{Cristina Pardo García} (\email{Cristina.Pardo-Garcia@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

Universidad de Valencia (España)
}
