% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momento.central.R
\name{momento.central}
\alias{momento.central}
\title{Momento central.}
\usage{
momento.central(x, orden)
}
\arguments{
\item{x}{Conjunto de datos. Puede ser un vector o un dataframe.}

\item{orden}{Es un valor numérico que representa el orden del momento central (orden = {1,2,3,4,...})}
}
\description{
Calcula los momentos centrales respecto de la media.
}
\references{
Esteban García, J. et al. (2005). Estadística descriptiva y nociones de probabilidad. Thomson.
}
\author{
\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

\strong{Olga Blasco-Blasco} (\email{olga.blasco@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

\strong{Rosario Martínez Verdú} (\email{rosario.martinez@uv.es}).
\emph{Economía Aplicada.}

\strong{Cristina Pardo García} (\email{cristina.pardo-garcia@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

Universidad de Valencia (España)
}
