% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion2.R
\name{my_fun_2}
\alias{my_fun_2}
\title{Esta función coge un df y calcula la media de cada columna}
\usage{
my_fun_2(x, y)
}
\arguments{
\item{df}{Un data.fram}
}
\value{
Un nuevo df con la media de cada columna del df original \code{df}
}
\description{
Esta función coge un df y calcula la media de cada columna
}
\examples{
my_fun_2(mtcars)
}
