% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.resumen.R
\name{print.resumen}
\alias{print.resumen}
\title{Método de impresión para objetos de clase "resumen"}
\usage{
\method{print}{resumen}(x, ...)
}
\arguments{
\item{x}{Objeto de clase "resumen"}

\item{...}{Argumentos adicionales}
}
\description{
Método de impresión para objetos de clase "resumen"
}
