% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.sets.R
\docType{data}
\name{rendimientos}
\alias{rendimientos}
\title{Datos simulados del rendimiento de dos acciones.}
\format{
Dataframe con 3000 observaciones de 3 variables.
\describe{
  \item{Dia}{dia}
  \item{X}{rendimiento mensual de la acción X, en euros}
  \item{Y}{rendimiento mensual de la acción Y, en euros}
}
}
\source{
Muestra simulada.
}
\usage{
data("rendimientos")
}
\description{
Datos simulados del rendimiento de dos acciones.
}
\author{
\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Quantitative Methods for Measuring Culture (MC2). Applied Economics.}

\strong{Rosario Martínez Verdú} (\email{rosario.martinez@uv.es}).
\emph{Economía Aplicada.}

\strong{Cristina Pardo García} (\email{cristina.pardo-garcia@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

Universidad de Valencia (España)
}
\keyword{datasets}
