% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.sets.R
\docType{data}
\name{salarios2018}
\alias{salarios2018}
\title{Data: Encuesta cuatrienal de estructura salarial (2018)}
\format{
Dataframe con 216726 observaciones de 10 variables.
\describe{
  \item{ID}{Identificador del encuestado.}
  \item{CODIGO.REGION}{1=NOROESTE,2=NORESTE,3=COMUNIDAD MADRID,4=CENTRO,5=ESTE,6=SUR,7=CANARIAS}
  \item{SEXO}{Sexo (1=hombre, 6=mujer)}
  \item{ESTUDIOS}{Nivel de estudios. 1=Menos que primaria,2=Primaria,3=Primera etapa secundaria,4=Segunda etapa secundaria,5=FP superior o similar,6=Diplomado o similar,7=Licenciados o similares y doctores}
  \item{TIPO.JORNADA}{Tipo de jornada laboral. 1=Tiempo completo,2=Tiempo parcial}
  \item{TIPO.CONTRATO}{Tipo de contrato laboral. 1=Indefinido,2=Duración determinada}
  \item{SALARIO.BRUTO.ANUAL}{Salario bruto anual}
  \item{SALARIO.ORDINARIO.ANUAL}{Salario ordinario anual}
  \item{PAGOS.EXTRA}{Pagos extraordinarios anuales}
  \item{FACTOR.ELEVACION}{Factor de elevación}
}
}
\source{
Instituto Nacional de Estadística \url{http://www.ine.es/}
}
\usage{
data("salarios2018")
}
\description{
Datos del Instituto Nacional de Estadística. Hay un total de 216,726 observaciones de 10 variables seleccionadas. Los datos han sido tratados siguiendo las instrucciones que el INE adjunta con los microdatos.
}
\author{
\strong{Vicente Coll-Serrano}.
\emph{Quantitative Methods for Measuring Culture (MC2). Applied Economics.}

\strong{Rosario Martínez Verdú}.
\emph{Economía Aplicada.}

\strong{Cristina Pardo-García}.
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

Facultad de Economía. Universidad de Valencia (España)
}
\keyword{datasets}
