% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.temporales.R
\name{series.temporales}
\alias{series.temporales}
\title{Series temporales.}
\usage{
series.temporales(x,
variable = NULL,
inicio_anual = 1,
periodo_inicio = 1,
frecuencia = 4,
orden = frecuencia)
}
\arguments{
<<<<<<< HEAD
\item{x}{sss}

\item{variable}{sss}

=======
>>>>>>> 6e2c4b7ae6e015d514381e59ce80402ad2abf3ae
\item{inicio_anual}{Año de inicio de la serie.}

\item{periodo_inicio}{Periodo de inicio de la serie.}

\item{frecuencia}{Periodificación de la serie.
Si anual, frecuencia = 1
Si semestral, frecuencia = 2
Si cuatrimestral, frecuencia = 3
Si trimestral, frecuencia = 4
Si bimestral, frecuencia = 6
Si mensual, frecuencia = 12
Si semanal, frecuencia = 52
Si diario, frecuencia = 360}

\item{orden}{Orden de la media móvil. Por defecto orden = frecuencia.}
<<<<<<< HEAD

\item{prediccion}{Orden de la media móvil. Por defecto orden = frecuencia.}

\item{grafico}{Orden de la media móvil. Por defecto orden = frecuencia.}

\item{exportar}{Orden de la media móvil. Por defecto orden = frecuencia.}
=======
>>>>>>> 6e2c4b7ae6e015d514381e59ce80402ad2abf3ae
}
\value{
Esta función devuelve la covarianza en un objeto de la clase \code{data.frame}.
}
\description{
Calcula la descomposición de las series temporales mediante el método de las medias móviles.
}
\details{

}
\note{

}
\references{
Esteban García, J. y otros. (2005). Estadística descriptiva y nociones de probabilidad. Paraninfo. ISBN: 9788497323741

Newbold, P, Carlson, W. y Thorne, B. (2019). Statistics for Business and Economics, Global Edition. Pearson. ISBN: 9781292315034

Murgui, J.S. y otros. (2002). Ejercicios de estadística Economía y Ciencias sociales. tirant lo blanch. ISBN: 9788484424673
}
\seealso{
\code{\link{varianza}}, \code{\link{desviacion}},\code{\link{matriz.covar}}
}
\author{
\strong{Vicente Coll-Serrano}.
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

\strong{Rosario Martínez Verdú}.
\emph{Economía Aplicada.}

Facultad de Economía. Universidad de Valencia (España)
}
