% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.sets.R
\docType{data}
\name{startup}
\alias{startup}
\title{Data: Datos de empresas emergentes (startups)}
\format{
Dataframe con 21 observaciones de 4 variables.
\describe{
  \item{gasto.desarrollo}{Gastos de investigación y desarrollo, en euros.}
  \item{gasto.marketing}{Gastos de marketing, en euros.}
  \item{gasto.gestion}{Gastos de administración, en euros.}
  \item{beneficio}{Beneficios, en euros.}
}
}
\source{
Muestra simulada.
}
\usage{
data("startup")
}
\description{
Datos simulados. Muestra de 21 empresas emergentes
}
\author{
\strong{Vicente Coll-Serrano}.
\emph{Quantitative Methods for Measuring Culture (MC2). Applied Economics.}

\strong{Rosario Martínez Verdú}.
\emph{Economía Aplicada.}

\strong{Cristina Pardo-García}.
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

Facultad de Economía. Universidad de Valencia (España)
}
\keyword{datasets}
