% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla.bidimensional.R
\name{tabla.bidimensional}
\alias{tabla.bidimensional}
\title{Tabla doble entrada.}
\usage{
tabla.bidimensional(x,
                           distribucion = c("cruzada","condicionada"),
                           frecuencias = c("absolutas","relativas"),
                           exportar = TRUE)
}
\arguments{
\item{x}{Conjunto de datos. Tiene que ser un dataframe (al menos dos variables, es decir, dos columnas).}

\item{distribucion}{Es un caracter. Por defecto se obtien la tabla cruzada (distribucion = "cruzada"). Para obtener las distribuciones condicionadas cambiar el argumento: distribucion = "condicionada".}

\item{frecuencias}{Es un carácter. Por defecto se obtienen las frecuencias absolutas ordinarias (frecuencias = "absolutas"). Para obtener las frecuencias relativas ordinarias cambiar el argumento: frecuencias = "relativas".}

\item{exportar}{Por defecto, los resultados se exportan a una hoja de cálculo Excel (exportar = TRUE).}
}
\description{
Calcula la tabla de frecuencias bidimensionales.
}
\references{
Esteban García, J. et al. (2005). Estadística descriptiva y nociones de probabilidad. Thomson.
}
\author{
\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

\strong{Olga Blasco-Blasco} (\email{olga.blasco@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

\strong{Rosario Martínez Verdú} (\email{rosario.martinez@uv.es}).
\emph{Economía Aplicada.}

\strong{Cristina Pardo García} (\email{cristina.pardo-garcia@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}
Universidad de Valencia (España)
}
