% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.sets.R
\docType{data}
\name{turistas2}
\alias{turistas2}
\title{Data: Turistas internacionales Comunidad Valenciana}
\format{
Dataframe con 80 observaciones de 2 variables.
\describe{
  \item{perido}{Periodo temporal.}
  \item{Turistas.internacionales}{Número de turistas con destino principal la Comunidad Valenciana}
}
}
\source{
Movimientos turísticos en fronteras. Frontur. Instituto de Estudios Turísticos (hasta septiembre de 2015) e INE (a partir de octubre de 2015)
}
\usage{
data("turistas")
}
\description{
Data: Turistas internacionales Comunidad Valenciana
}
\keyword{datasets}
