% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.sets.R
\docType{data}
\name{viajes_vendidos}
\alias{viajes_vendidos}
\title{Data: Viajes vendidos}
\format{
Dataframe con 5 observaciones de 3 variables.
\describe{
  \item{Número.de.viajes.vendidos}{Número de viajes perdidos.}
  \item{Empleados}{Número de empleados}
  \item{Ni}{Frecuencia absoluta acumulada del número de empleados}
}
}
\usage{
data("viajes_vendidos")
}
\description{
Datos de 5 observaciones que se utilizan en los ejemplos de (1) media, mediana y moda, (2) cuantiles, (3) varianza, desviación típica y coeficiente de variación, (4) medidas de forma y momento central y (5) resumen de descriptivos
}
\keyword{datasets}
