% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribuciones.probabilidad.R
\name{distribuciones.probabilidad}
\alias{distribuciones.probabilidad}
\title{Distribuciones de probabilidad.}
\usage{
distribuciones.probabilidad()
}
\description{
Aplicación interactiva donde se representa las principales distribuciones de probabilidad unidimensionales: Binomial, Poisson, Uniforme, Exponencial y Normal.
}
\author{
\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

\strong{Rosario Martínez Verdú} (\email{rosario.martinez@uv.es}).
\emph{Economía Aplicada.}

\strong{Cristina Pardo García} (\email{cristina.pardo-garcia@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

Facultad de Economía. Universidad de Valencia (España)
}
