% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nivel.confianza.R
\encoding{UTF-8}
\name{nivel.confianza}
\alias{nivel.confianza}
\title{Nivel de confianza.}
\usage{
nivel.confianza(min.pob,
max.pob,
muestras = 200,
n = 100,
confianza = 0.95,
grafico = TRUE,
exportar = FALSE,
replicar = TRUE)
}
\arguments{
\item{min.pob}{Es un valor numérico que indica el valor mínimo poblacional. Por defecto min.pob = 2000}

\item{max.pob}{Es un valor numérico que indica el valor máximo poblacional. Por defecto max.pob = 45000}

\item{muestras}{Es un valor numérico entre 50 y 10000 que indica el número de muestras que se extraen sin reemplazamiento de la población. Por defecto muestras = 200}

\item{n}{Es un valor numérico entre 25 y 2000 que indica el tamaño de la muestra. Por defecto n = 100.sumidos en una distribución de frecuencias, debe indicarse la columna que representa los valores de la variable y la columna con las frecuencias o pesos.}

\item{confianza}{Es un valor numérico entre 0 y 1. Indica el nivel de confianza. Por defecto, confianza = 0.95 (95 por ciento)}

\item{grafico}{Si grafico = TRUE se representan los intervalos de confianza de las muestras seleccionadas y la media poblacional.}

\item{exportar}{Para exportar los resultados a una hoja de cálculo Excel (exportar = TRUE).}

\item{replicar}{Es un valor lógico. Si replicar=TRUE se fija una semilla para que los resultados sean reproducibles. Si replicar=FALSE los resultados serán aleatorios y cambiarán en cada realización.}
}
\description{
Esta función simula una población de tamaño 100,000 de la que se extraen diversas muestras y construye los correspondientes intervalos de confianzas. El objetivo es transmitir el concepto de nivel de confianza.
}
\note{
Si se seleccionan 10000 muestras de tamaño 2000, el tiempo estimado de ejecución es de 9 minutos.
}
\references{
Casas José M. () Inferencia estadística. Editoral: Centro de estudios Ramón Areces, S.A. ISBN: 848004263-X

Esteban García, J. et al. (2008). Curso básico de inferencia estadística. ReproExprés, SL. ISBN: 8493036595.

Murgui, J.S. y otros. (2002). Ejercicios de estadística Economía y Ciencias sociales. tirant lo blanch. ISBN: 9788484424673

Newbold, P, Carlson, W. y Thorne, B. (2019). Statistics for Business and Economics, Global Edition. Pearson. ISBN: 9781292315034
}
\seealso{
\code{\link{ic.media}}
}
\author{
\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

\strong{Rosario Martínez Verdú} (\email{rosario.martinez@uv.es}).
\emph{Economía Aplicada.}

\strong{Cristina Pardo García} (\email{cristina.pardo-garcia@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

Universidad de Valencia (España)
}
