% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla.frecuencias.R
\name{tabla.frecuencias}
\alias{tabla.frecuencias}
\title{Tabla de frecuencias.}
\usage{
tabla.frecuencias(x,
eliminar.na = TRUE,
exportar = FALSE)
}
\arguments{
\item{x}{Conjunto de datos. Puede ser un vector o un dataframe. Si el dataframe tiene más de una variable, solicitará al usuario que idenfique el nombre de la variable para la que se quiere calcular la tabla de frecuencias.}

\item{eliminar.na}{Valor lógico. Por defecto eliminar.na = TRUE. Si se quiere obtener la tabla de frecuencias con NAs, cambiar el argumento a FALSE.}

\item{exportar}{Para exportar los resultados a una hoja de cálculo Excel (exportar = TRUE).}
}
\description{
Esta función presenta la distribución de frecuencias de una variable.
}
\references{
Esteban García, J. y otros. (2005). Estadística descriptiva y nociones de probabilidad. Paraninfo. ISBN: 9788497323741

Newbold, P, Carlson, W. y Thorne, B. (2019). Statistics for Business and Economics, Global Edition. Pearson. ISBN: 9781292315034

Murgui, J.S. y otros. (2002). Ejercicios de estadística Economía y Ciencias sociales. tirant lo blanch. ISBN: 9788484424673
}
\author{
\strong{Vicente Coll-Serrano} (\email{vicente.coll@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

\strong{Rosario Martínez Verdú} (\email{rosario.martinez@uv.es}).
\emph{Economía Aplicada.}

\strong{Cristina Pardo García} (\email{cristina.pardo-garcia@uv.es}).
\emph{Métodos Cuantitativos para la Medición de la Cultura (MC2). Economía Aplicada.}

Universidad de Valencia (España)
}
