% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.sets.R
\docType{data}
\name{turistas}
\alias{turistas}
\title{Data: Turistas por paises (WTO)}
\format{
Dataframe con 130 observaciones de 3 variables.
\describe{
  \item{País}{País de destino.}
  \item{Llegadas.turistas}{Número de llegada de turistas en 2017, en miles.}
  \item{Gasto.viajes}{Gasto en viajes en 2017, en millones de USD.}
}
}
\source{
World Tourism Organization (2019).
}
\usage{
data("turistas")
}
\description{
Datos de World Tourism Organization.
}
\keyword{datasets}
