% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.sets.R
\docType{data}
\name{turistas}
\alias{turistas}
\title{Data: turistas}
\format{
Dataframe con 80 observaciones de 2 variables.
\describe{
  \item{perido}{Periodo temporal.}
  \item{Turistas.internacionales}{Número de turistas internacionales}
}
}
\source{
Instituto Nacional de Estadística \url{http://www.ine.es/}
}
\usage{
data("turistas")
}
\description{
Datos del Instituto Nacional de Estadística. Hay un total de 216,726 observaciones de 10 variables seleccionadas. Los datos han sido tratados siguiendo las instrucciones que el INE adjunta con los microdatos.
}
\keyword{datasets}
